requirements.txt
pytest
pytest-playwright
pytest-html
python-dotenv
